[DEV-67] 1.8 회원가입 api

+++ src/main/java/community/mingle/api/domain/member/repository/UniversityRepository.java
@@ -4,9 +4,8 @@ package community.mingle.api.domain.member.repository;
 import community.mingle.api.domain.member.entity.University;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import java.util.List;
 
 @Repository
-public interface UniversityRepository extends JpaRepository<University, Long> {
+public interface UniversityRepository extends JpaRepository<University, Integer> {
 
 }
